# openapi.yaml
openapi: 3.0.3
info:
  title: Northbound - Order Creation SaaS API
  description: |
    A SaaS API for generating and managing standardized UBL 2.1 Order documents.
    This service facilitates the procurement process between Buyer and Seller parties.

    **Versioning:** Create/list/get/xml use `/orders` and `/v1/orders`. **Order change** is implemented
    under **`/v2/orders/{orderID}/change`** (Sprint 3). Older `PUT /orders/{orderID}/change` returns 501
    with guidance to use v2.
  version: "1.1.0"

servers:
  - url: https://northbound-w6b3.onrender.com
    description: Production (Render)
  - url: http://localhost:3000
    description: Local development

tags:
  - name: Health
  - name: Auth
  - name: Orders
  - name: OrdersV2

security:
  - bearerAuth: []

paths:
  /health:
    get:
      tags: [Health]
      summary: Service Health Check
      description: Returns the aliveness status and basic runtime info for the service.
      security: []
      responses:
        "200":
          description: Service is operational
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /auth/login:
    post:
      tags: [Auth]
      summary: User Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Missing or invalid fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"

  /auth/register:
    post:
      tags: [Auth]
      summary: User Registration
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Missing or invalid fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"

  /orders:
    post:
      tags: [Orders]
      summary: Generate and store UBL Order
      description: |
        Creates a new Order from structured JSON input, generates a UBL 2.1 Order XML document,
        stores it, and returns the created order identifier and XML.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateRequest"
      responses:
        "201":
          description: Order created and UBL XML generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedOrderResponse"
        "400":
          description: Missing mandatory fields or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "422":
          description: UBL schema validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"
    get:
      tags: [Orders]
      summary: List Orders
      description: Returns a list of orders.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Optional filter by order status
      responses:
        "200":
          description: List of order rows from storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"

  /orders/generate:
    post:
      tags: [Orders]
      summary: Generate and store UBL Order (/generate alias)
      description: Generates and stores an order from structured JSON input.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateRequest"
      responses:
        "201":
          description: Order created and UBL XML generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedOrderResponse"
        "400":
          description: Missing mandatory fields or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "422":
          description: UBL schema validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"

  /orders/{orderID}:
    get:
      tags: [Orders]
      summary: Retrieve Order by ID
      description: Returns the stored order row for a given ID.
      parameters:
        - name: orderID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderGetResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"

  /orders/{orderID}/xml:
    get:
      tags: [Orders]
      summary: Retrieve UBL Order XML by ID
      description: Returns the UBL 2.1 Order XML document for a given order ID.
      parameters:
        - name: orderID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: UBL XML document
          content:
            application/xml:
              schema:
                type: string
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"

  /orders/{orderID}/change:
    put:
      tags: [Orders]
      summary: Change / replace an existing order (v1)
      description: |
        Order change is not implemented on this API version.
        Use `PUT /v2/orders/{orderID}/change` instead.
      parameters:
        - name: orderID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateRequest"
      responses:
        "501":
          description: Order change not available; use v2
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"

  /orders/{orderID}/cancel:
    post:
      tags: [Orders]
      summary: Cancel an order (not implemented in v1)
      description: Not implemented in this API version. Use v2 when available.
      parameters:
        - name: orderID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "501":
          description: Order cancel not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"

  /orders/{orderID}/response:
    post:
      tags: [Orders]
      summary: Submit an order response (not implemented in v1)
      description: Not implemented in this API version. Use v2 when available.
      parameters:
        - name: orderID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "501":
          description: Order response not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"

  /orders/{orderID}/detail:
    patch:
      tags: [Orders]
      summary: Update order detail (not implemented in v1)
      description: Not implemented in this API version. Use v2 when available.
      parameters:
        - name: orderID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "501":
          description: Order detail update not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"

  /v2/orders/{orderID}/cancel:
    post:
      tags: [OrdersV2]
      summary: Cancel an order (v2)
      description: |
        **Sprint 3 / API v2.** Cancels an existing order (deletes the order and associated line items).
      parameters:
        - name: orderID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order cancelled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelledOrderResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"
        "500":
          description: Database or configuration error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"

  /v2/orders/{orderID}/change:
    put:
      tags: [OrdersV2]
      summary: Change / replace an existing order (v2)
      description: |
        **Sprint 3 / API v2.** Replaces the order with a full new payload (same JSON shape as `POST /orders`),
        updates parties and line items in the database, and returns regenerated UBL XML. The order **id** does not change.

        Use this endpoint instead of `PUT /orders/{orderID}/change`, which returns 501.
      parameters:
        - name: orderID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateRequest"
            examples:
              fullReplace:
                summary: Full order body (same as create)
                value:
                  buyer:
                    external_id: BUYER-001
                    name: Example Buyer Pty Ltd
                  seller:
                    external_id: SELLER-001
                    name: Example Seller Pty Ltd
                  currency: AUD
                  issue_date: "2025-03-15"
                  order_note: Updated shipment
                  order_lines:
                    - line_id: "1"
                      description: Wireless Keyboard
                      quantity: 3
                      unit_price: 49.99
                      unit_code: EA
      responses:
        "200":
          description: Order updated; new UBL XML returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatedOrderResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"
        "500":
          description: Database or configuration error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Supply as: Authorization: Bearer <token>'

  schemas:
    StandardResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully.
        data:
          type: object
          nullable: true
          description: Payload specific to the endpoint.
        error:
          type: object
          nullable: true
          description: Error information if success is false.

    StandardErrorResponse:
      allOf:
        - $ref: "#/components/schemas/StandardResponse"
        - type: object
          properties:
            success:
              example: false
            data:
              nullable: true
              example: null
            error:
              $ref: "#/components/schemas/ErrorDetail"

    ValidationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/StandardErrorResponse"
        - type: object
          properties:
            error:
              allOf:
                - $ref: "#/components/schemas/ErrorDetail"
                - type: object
                  properties:
                    validationErrors:
                      type: array
                      items:
                        $ref: "#/components/schemas/ValidationError"

    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: One or more fields are invalid.

    ValidationError:
      type: object
      properties:
        field:
          type: string
          example: buyer.external_id
        message:
          type: string
          example: buyer.external_id is required

    HealthStatus:
      type: object
      properties:
        status:
          type: string
          example: UP
        uptime:
          type: number
          example: 0
        version:
          type: string
          example: "1.0.0"

    HealthResponse:
      allOf:
        - $ref: "#/components/schemas/StandardResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/HealthStatus"

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    RegisterRequest:
      type: object
      required: [email, password, passwordConfirm]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        passwordConfirm:
          type: string

    AuthToken:
      type: object
      properties:
        userID:
          type: integer
          example: 123
        token:
          type: string

    AuthResponse:
      allOf:
        - $ref: "#/components/schemas/StandardResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/AuthToken"

    OrderCreateRequest:
      type: object
      required: [buyer, seller, currency, issue_date, order_lines]
      properties:
        buyer:
          $ref: "#/components/schemas/Party"
        seller:
          $ref: "#/components/schemas/Party"
        currency:
          type: string
          description: ISO 4217 currency code
          example: AUD
        issue_date:
          type: string
          format: date
          description: Order date (YYYY-MM-DD)
          example: "2025-03-10"
        order_note:
          type: string
          description: Optional note for the order
        order_lines:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderLine"

    Party:
      type: object
      required: [external_id, name]
      properties:
        external_id:
          type: string
          example: BUYER-001
        name:
          type: string
          example: Example Buyer Pty Ltd
        email:
          type: string
          format: email
        street:
          type: string
        city:
          type: string
        country:
          type: string
          example: AU
        postal_code:
          type: string

    OrderLine:
      type: object
      required: [line_id, description, quantity, unit_price]
      properties:
        line_id:
          type: string
          example: "1"
        description:
          type: string
          example: Wireless Keyboard
        quantity:
          type: number
          minimum: 0
          example: 2
        unit_price:
          type: number
          minimum: 0
          example: 49.99
        unit_code:
          type: string
          example: EA

    CreatedOrder:
      type: object
      properties:
        orderID:
          type: string
          format: uuid
        ubl_xml:
          type: string
          description: Generated UBL 2.1 Order XML document

    CreatedOrderResponse:
      allOf:
        - $ref: "#/components/schemas/StandardResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/CreatedOrder"

    UpdatedOrderResponse:
      allOf:
        - $ref: "#/components/schemas/StandardResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/CreatedOrder"
              description: Same shape as create — orderID unchanged, ubl_xml regenerated

    CancelledOrder:
      type: object
      properties:
        orderID:
          type: string
          format: uuid
      required: [orderID]

    CancelledOrderResponse:
      allOf:
        - $ref: "#/components/schemas/StandardResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/CancelledOrder"

    OrderSummary:
      type: object
      additionalProperties: true

    OrderObject:
      type: object
      additionalProperties: true

    OrderListResponse:
      allOf:
        - $ref: "#/components/schemas/StandardResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/OrderSummary"

    OrderGetResponse:
      allOf:
        - $ref: "#/components/schemas/StandardResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/OrderObject"